requirements.txt
attrs==20.3.0
biopython==1.78
certifi==2020.12.5
chardet==4.0.0
colorlog==4.8.0
cycler==0.10.0
decorator==4.4.2
dna-features-viewer==3.0.3
idna==2.10
importlib-metadata==4.0.1
inflection==0.3.1
jsonschema==3.2.0
kiwisolver==1.3.1
matplotlib==3.3.4
nest-asyncio==1.5.1
networkx==2.5.1
numpy==1.19.5
Pillow==8.2.0
py==1.10.0
pydent==1.0.8
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
requests==2.25.1
retry==0.9.2
six==1.15.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
